feat: Login First error if user is not logged in

+++ src/services/noteServices.js
@@ -13,25 +13,29 @@ const saveNoteHandler = async(
     encodedToken
 ) => {
     e.preventDefault()
-    if (allFieldsAreFilled(note)) {
-        try {
-            const response = await axios.post(
-                '/api/notes', {
-                    note: {...note, createdOn: getCurrentDate() },
-                }, {
-                    headers: { authorization: encodedToken },
-                }
-            )
-            setNotes(response.data.notes)
-            setShowPalette(false)
-            response.status === 201 && noteDispatch({ type: 'RESET' })
-            successToast('You have successfully saved the note!')
-        } catch (err) {
-            console.log(err)
-            errorToast('Something went wrong, Please try again!')
+
+    if (encodedToken) {
+        if (allFieldsAreFilled(note)) {
+            try {
+                const response = await axios.post(
+                    '/api/notes', {
+                        note: {...note, createdOn: getCurrentDate() },
+                    }, {
+                        headers: { authorization: encodedToken },
+                    }
+                )
+                setNotes(response.data.notes)
+                setShowPalette(false)
+                response.status === 201 && noteDispatch({ type: 'RESET' })
+                successToast('You have successfully saved the note!')
+            } catch (err) {
+                errorToast('Something went wrong, Please try again!')
+            }
+        } else {
+            errorToast('Please fill all the fields')
         }
     } else {
-        errorToast('Please fill all the fields')
+        errorToast('Please Login First!')
     }
 }
 
@@ -44,19 +48,23 @@ const editNoteHandler = async(
     encodedToken
 ) => {
     e.preventDefault()
-    try {
-        const response = await axios.post(
-            `/api/notes/${_id}`, {
-                note: editNote,
-            }, {
-                headers: { authorization: encodedToken },
-            }
-        )
-        successToast('You have Successfully Edited the note')
-        setOpenEdit(false)
-        setNotes(response.data.notes)
-    } catch (err) {
-        errorToast('Could not Edit the note, please try again!')
+    if (encodedToken) {
+        try {
+            const response = await axios.post(
+                `/api/notes/${_id}`, {
+                    note: editNote,
+                }, {
+                    headers: { authorization: encodedToken },
+                }
+            )
+            successToast('You have Successfully Edited the note')
+            setOpenEdit(false)
+            setNotes(response.data.notes)
+        } catch (err) {
+            errorToast('Could not Edit the note, please try again!')
+        }
+    } else {
+        errorToast('Please Login First!')
     }
 }
 
@@ -70,15 +78,19 @@ const deleteNoteHandler = async(
     encodedToken
 ) => {
     e.preventDefault()
-    try {
-        const response = await axios.delete(`/api/notes/${_id}`, {
-            headers: { authorization: encodedToken },
-        })
-        setNotes(response.data.notes)
-        setTrash([...trash, note])
-        successToast('Successfully deleted the note')
-    } catch (err) {
-        errorToast('Could not delete the note, please try again!')
+    if (encodedToken) {
+        try {
+            const response = await axios.delete(`/api/notes/${_id}`, {
+                headers: { authorization: encodedToken },
+            })
+            setNotes(response.data.notes)
+            setTrash([...trash, note])
+            successToast('Successfully deleted the note')
+        } catch (err) {
+            errorToast('Could not delete the note, please try again!')
+        }
+    } else {
+        errorToast('Please Login First!')
     }
 }
 
@@ -92,27 +104,35 @@ const deleteArchiveNoteHandler = async(
     encodedToken
 ) => {
     e.preventDefault()
-    try {
-        const response = await axios.delete(`/api/archives/delete/${_id}`, {
-            headers: { authorization: encodedToken },
-        })
-        setArchiveNotes(response.data.archives)
-        setTrash([...trash, note])
-        successToast('Successfully deleted the note')
-    } catch (err) {
-        errorToast('Could not delete the note, please try again!')
+    if (encodedToken) {
+        try {
+            const response = await axios.delete(`/api/archives/delete/${_id}`, {
+                headers: { authorization: encodedToken },
+            })
+            setArchiveNotes(response.data.archives)
+            setTrash([...trash, note])
+            successToast('Successfully deleted the note')
+        } catch (err) {
+            errorToast('Could not delete the note, please try again!')
+        }
+    } else {
+        errorToast('Please Login First!')
     }
 }
 
 const archiveNoteHandler = async(_id, note, setNotes, encodedToken) => {
-    try {
-        const response = await axios.post(
-            `/api/notes/archives/${_id}`, { note }, { headers: { authorization: encodedToken } }
-        )
-        setNotes(response.data.notes)
-        successToast('Note has been successfully archived')
-    } catch (error) {
-        errorToast('Note was not archived, Please try again!')
+    if (encodedToken) {
+        try {
+            const response = await axios.post(
+                `/api/notes/archives/${_id}`, { note }, { headers: { authorization: encodedToken } }
+            )
+            setNotes(response.data.notes)
+            successToast('Note has been successfully archived')
+        } catch (error) {
+            errorToast('Note was not archived, Please try again!')
+        }
+    } else {
+        errorToast('Please Login First!')
     }
 }
 
@@ -123,17 +143,21 @@ const restoreNoteHandler = async(
     encodedToken,
     setArchiveNotes
 ) => {
-    try {
-        const response = await axios.post(
-            `/api/archives/restore/${_id}`, {}, {
-                headers: { authorization: encodedToken },
-            }
-        )
-        successToast('You have Successfully restored the note')
-        setNotes(response.data.notes)
-        setArchiveNotes(response.data.archives)
-    } catch (err) {
-        errorToast('Could not restore the note, please try again!')
+    if (encodedToken) {
+        try {
+            const response = await axios.post(
+                `/api/archives/restore/${_id}`, {}, {
+                    headers: { authorization: encodedToken },
+                }
+            )
+            successToast('You have Successfully restored the note')
+            setNotes(response.data.notes)
+            setArchiveNotes(response.data.archives)
+        } catch (err) {
+            errorToast('Could not restore the note, please try again!')
+        }
+    } else {
+        errorToast('Please Login First!')
     }
 }
